feat: next update and docs improvements
