+++ docs/src/components/FeatureCards/FeatureCard.tsx
@@ -1,4 +1,4 @@
-import { morfeo } from '@morfeo/web';
+import { morfeo } from '@morfeo/react';
 import {
   DetailedHTMLProps,
   HTMLAttributes,
