+++ docs/src/components/Footer/LinkBlocks.tsx
@@ -1,4 +1,4 @@
-import { morfeo } from '@morfeo/web';
+import { morfeo } from '@morfeo/react';
 import { FooterLinkBlock } from './types';
 
 type Props = {
