+++ docs/src/components/Footer/Social.tsx
@@ -1,4 +1,4 @@
-import { morfeo } from '@morfeo/web';
+import { morfeo } from '@morfeo/react';
 
 const classes = morfeo.css({
   container: {
