+++ docs/src/components/HowItWorks/Code.mdx
@@ -1,5 +1,5 @@
 ```tsx showLineNumbers {1, 3-8} /classes('button')/
-import { morfeo } from '@morfeo/web';
+import { morfeo } from '@morfeo/react';
 
 const classes = morfeo.css({
   button: {
