+++ docs/src/components/HowItWorks/ThemeSnippet.mdx
@@ -1,5 +1,5 @@
 ```tsx showLineNumbers
-import { theme } from '@morfeo/web';
+import { theme } from '@morfeo/react';
 
 theme.set({
   colors: {
