+++ docs/src/components/WebExtensionSection/WebExtensionSection.tsx
@@ -1,4 +1,4 @@
-import { morfeo } from '@morfeo/web';
+import { morfeo } from '@morfeo/react';
 import Image from 'next/image';
 import Link from 'next/link';
 import { WebExtensionStepProps } from './types';
