+++ docs/src/pages/_app.tsx
@@ -1,7 +1,6 @@
 import { AppProps } from 'next/app';
 import '@/styles/globals.css';
 import '@/styles/morfeo.css';
-import '@morfeo/react';
 import '../../morfeo.theme';
 
 export default function RootLayout({ Component, pageProps }: AppProps) {
