+++ docs/src/pages/index.mdx
@@ -1,3 +1,7 @@
+---
+title: Ⓜ️ Morfeo - The design library
+---
+
 import { Hero } from '@/components/Hero';
 import { FeatureCards } from '@/components/FeatureCards';
 import { HowItWorks } from '@/components/HowItWorks';
