+++ examples/next/src/app/layout.tsx
@@ -1,4 +1,6 @@
 import './globals.css';
+import '../../morfeo.theme';
+import '../styles/morfeo.css';
 
 type LayoutProps = {
   children: React.ReactNode;
