+++ package.json
@@ -24,7 +24,6 @@
   "dependencies": {
     "@changesets/cli": "2.26.2",
     "csstype": "3.1.2",
-    "next": "13.5.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-native": "0.71.7"
