+++ packages/web/src/utils/expandStyles.ts
@@ -1,5 +1,6 @@
 import { deepMerge, isDefaultObject } from '@morfeo/utils';
 import { Style, morfeo } from '@morfeo/core';
+import { isMultiThemeValue } from './multiTheming';
 
 type Callback = (value: Style) => Style;
 
@@ -46,7 +47,7 @@ export function expandStyles(
                 restStyle,
                 pipe(
                   getContext,
-                  (value: Style) => ({ [curr]: value } as Style),
+                  (value: Style) => ({ [curr]: value }) as Style,
                 ),
               ),
             },
@@ -57,7 +58,7 @@ export function expandStyles(
           ...acc,
           [curr]: traverse(
             currentStyle,
-            pipe(getContext, (value: Style) => ({ [curr]: value } as Style)),
+            pipe(getContext, (value: Style) => ({ [curr]: value })),
           ),
         };
       }
